minor msvc fixes, more RIO_NO_GLFW_CALLS cases
* reduce rio_Math.h constants since msvc interpreted them as doubles
* add RIO_NO_TEXTURE2D_FILE_CTOR, and extend RIO_NO_GLFW_CALLS

... sooo you should be able to build FFL with libsupc++ with:

gcc -Os -IninTexUtils/include -I/home/arian/Downloads/rio/include
-Iffl/include -DRIO_USE_GLEW (find ffl/src/ -name '*.c' -or -name
'*.cpp') -Iinclude -o /dev/shm/ffl_testing_2_debug64_supc -std=c++17 -lz
-lGLEW -DRIO_DEBUG -DRIO_NO_GLFW_CALLS -lGL -DNDEBUG
-DRIO_NO_CONTROLLERS_WIN -DRIO_NO_CLIP_CONTROL -lsupc++ -lm
-DFFL_NO_DATABASE_FILE -DFFL_NO_MIDDLE_DB -DFFL_NO_DATABASE_DEFAULT
-DFFL_NO_DATABASE_RANDOM -DFFL_NO_FS -DFFL_NO_NINTEXUTILS
-DRIO_NO_TEXTURE2D_FILE_CTOR ~/Downloads/rio/src/gpu/rio_Texture.cpp
~/Downloads/rio/src/gpu/win/rio_Texture2DUtilWin.cpp
~/Downloads/rio/src/gpu/win/rio_TextureWin.cpp
~/Downloads/rio/src/gfx/rio_Projection.cpp
~/Downloads/rio/src/gpu/rio_RenderState.cpp
~/Downloads/rio/src/gpu/win/rio_RenderStateWin.cpp
~/Downloads/rio/src/gpu/rio_RenderBuffer.cpp
~/Downloads/rio/src/math/rio_Matrix.cpp
~/Downloads/rio/src/gpu/rio_RenderTarget.cpp
~/Downloads/rio/src/gfx/win/rio_GraphicsWin.cpp -ffunction-sections
-fdata-sections -fvisibility=hidden -fPIC

+++ src/controller/win/rio_KeyboardMouseDeviceWin.cpp
@@ -27,7 +27,7 @@ void KeyboardMouseDevice::calc()
 
     f64 pos_x, pos_y;
     glfwGetCursorPos(Window::getWindowInner(), &pos_x, &pos_y);
-    mCursorPos.set(pos_x, pos_y);
+    mCursorPos.set(static_cast<f32>(pos_x), static_cast<f32>(pos_y));
 }
 
 }
